Conflicts:
	Cyberduck.xcodeproj/project.pbxproj

Former-commit-id: 8d90d3648e69c5dc0a53e08ae6d54a3facab11ad

+++ spectra/src/main/java/ch/cyberduck/core/spectra/SpectraDeleteFeature.java
@@ -20,11 +20,13 @@ import ch.cyberduck.core.LoginCallback;
 import ch.cyberduck.core.Path;
 import ch.cyberduck.core.PathContainerService;
 import ch.cyberduck.core.exception.BackgroundException;
+import ch.cyberduck.core.s3.S3DisabledMultipartService;
 import ch.cyberduck.core.s3.S3MultipleDeleteFeature;
 import ch.cyberduck.core.s3.S3PathContainerService;
 
 import java.io.IOException;
 import java.security.SignatureException;
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
@@ -40,22 +42,26 @@ public class SpectraDeleteFeature extends S3MultipleDeleteFeature {
     private final SpectraSession session;
 
     public SpectraDeleteFeature(final SpectraSession session) {
-        super(session);
+        super(session, new S3DisabledMultipartService());
         this.session = session;
     }
 
     @Override
     public void delete(final List<Path> files, final LoginCallback prompt, final Callback callback) throws BackgroundException {
         try {
-            for(Iterator<Path> iter = files.iterator(); iter.hasNext(); ) {
+            final ArrayList<Path> filtered = new ArrayList<Path>(files);
+            for(Iterator<Path> iter = filtered.iterator(); iter.hasNext(); ) {
                 final Path file = iter.next();
+                if(file.isVolume()) {
+                    continue;
+                }
                 if(file.isDirectory()) {
                     final Ds3Client client = new SpectraClientBuilder().wrap(session);
                     client.deleteFolder(new DeleteFolderRequest(containerService.getContainer(file).getName(), containerService.getKey(file)));
                     iter.remove();
                 }
             }
-            super.delete(files, prompt, callback);
+            super.delete(filtered, prompt, callback);
         }
         catch(FailedRequestException e) {
             throw new SpectraExceptionMappingService().map(e);
